Refactor audio.py into smaller modules
