+++ lhotse/audio/__init__.py
@@ -1,0 +1,20 @@
+from .backend import (
+    get_current_audio_backend,
+    get_default_audio_backend,
+    get_ffmpeg_torchaudio_info_enabled,
+    info,
+    read_audio,
+    set_current_audio_backend,
+    set_ffmpeg_torchaudio_info_enabled,
+)
+from .recording import Recording
+from .recording_set import RecordingSet
+from .source import AudioSource
+from .utils import (
+    AudioLoadingError,
+    DurationMismatchError,
+    get_audio_duration_mismatch_tolerance,
+    null_result_on_audio_loading_error,
+    set_audio_duration_mismatch_tolerance,
+    suppress_audio_loading_errors,
+)
