+++ lhotse/dataset/unsupervised.py
@@ -6,7 +6,8 @@ import torch
 from torch.utils.data import IterableDataset
 
 from lhotse import RecordingSet, Seconds, compute_num_samples, validate
-from lhotse.audio import suppress_audio_loading_errors, torchaudio_supports_ffmpeg
+from lhotse.audio.backend import torchaudio_supports_ffmpeg
+from lhotse.audio.utils import suppress_audio_loading_errors
 from lhotse.augmentation import AugmentFn
 from lhotse.cut import CutSet
 from lhotse.dataset.collation import collate_audio, collate_features, collate_matrices
