+++ lhotse/features/librosa_fbank.py
@@ -120,7 +120,9 @@ def logmelfilterbank(
 
     fmin = 0 if fmin is None else fmin
     fmax = sampling_rate / 2 if fmax is None else fmax
-    mel_basis = librosa.filters.mel(sampling_rate, fft_size, num_mel_bins, fmin, fmax)
+    mel_basis = librosa.filters.mel(
+        sr=sampling_rate, n_fft=fft_size, n_mels=num_mel_bins, fmin=fmin, fmax=fmax
+    )
 
     feats = np.log10(np.maximum(eps, np.dot(spc, mel_basis.T)))
