+++ test/audio/test_audio_reads.py
@@ -7,7 +7,7 @@ import torchaudio
 
 import lhotse
 from lhotse import AudioSource, Recording
-from lhotse.audio import read_opus_ffmpeg, read_opus_torchaudio, torchaudio_load
+from lhotse.audio.backend import read_opus_ffmpeg, read_opus_torchaudio, torchaudio_load
 
 
 @pytest.mark.parametrize(
@@ -169,7 +169,6 @@ def test_command_audio_caching_enabled_works():
 
         # Read the audio -- should be equal to noise1.
         audio = audio_source.load_audio()
-        audio = np.expand_dims(audio, axis=0)
         np.testing.assert_allclose(audio, noise1, atol=3e-5)
 
         # Save noise2 to the same location.
@@ -178,7 +177,6 @@ def test_command_audio_caching_enabled_works():
         # Read the audio -- should *still* be equal to noise1,
         # because reading from this path was cached before.
         audio = audio_source.load_audio()
-        audio = np.expand_dims(audio, axis=0)
         np.testing.assert_allclose(audio, noise1, atol=3e-5)
 
 
@@ -201,7 +199,6 @@ def test_command_audio_caching_disabled_works():
 
         # Read the audio -- should be equal to noise1.
         audio = audio_source.load_audio()
-        audio = np.expand_dims(audio, axis=0)
         np.testing.assert_allclose(audio, noise1, atol=3e-5)
 
         # Save noise2 to the same location.
@@ -210,7 +207,6 @@ def test_command_audio_caching_disabled_works():
         # Read the audio -- should be equal to noise2,
         # and the caching is ignored (doesn't happen).
         audio = audio_source.load_audio()
-        audio = np.expand_dims(audio, axis=0)
         np.testing.assert_allclose(audio, noise2, atol=3e-5)
