+++ test/cut/test_multi_cut_augmentation.py
@@ -1,11 +1,7 @@
 import numpy as np
 import pytest
-import torch
 
-from lhotse import AudioSource, CutSet, MultiCut, Recording, SupervisionSegment
-from lhotse.audio import RecordingSet
-from lhotse.cut import PaddingCut
-from lhotse.utils import fastcopy
+from lhotse import MultiCut, Recording, SupervisionSegment
 
 
 @pytest.fixture
@@ -198,7 +194,6 @@ def test_cut_reverb_rir(
     rir = request.getfixturevalue(rir)
     cut = cut_with_supervision
     cut_rvb = cut.reverb_rir(rir, rir_channels=rir_channels)
-    print(cut_rvb.channel)
     assert cut_rvb.start == cut.start
     assert cut_rvb.duration == cut.duration
     assert cut_rvb.end == cut.end
